Add files via upload
This is my Project for COMP165 course. This project is about a Stock Website which gives a brief description about the stocks on the main webpage. Then there are two different buttons below the description of the stocks. 
The first button takes to the another webpage which shows us the prices of some of the popular stocks.
The second button takes us to another webpage which shows the current chart for any stock you want to search for.